Moved the range utility
