+++ src/utils/index.ts
@@ -19,3 +19,4 @@
 export { countdown } from './countdown';
 export { timeout } from './timeout';
 export { wait } from './wait';
+export { getRange } from './range';
